event_writer/requirements.txt
uvicorn >= 0.20.0
sqlalchemy >= 1.4.44
kafka-python == 2.0.2